Add picture URL to LinkedIn user info endpoint default fields
Extended the list of fields included on the LinkedIn basic profile scope
endpoint to include the picture URL also.  It is nice having this on the
user object to parse out in existing authentication providers such as
Google and Yahoo and having it on LinkedIn is also nice for the sake of
consistency across the spectrum of supported options.  Similar to the
Google implementation this is not promoted to a top level authenticated
context property but just hangs off of the user object graph in the
event somebody needs it to not incur the small overhead of parsing out
each time.